Fix initialisation order and enable warnings
Class members are initialised in the order they are declared in the
class. The order you choose/attempt to initialise them in the
initialisation list is completely ignored by the language, and the
compiler will still obey the initial order of declaration.

Therefore, before this fix, the m_listCount, m_valueMax, and
m_valueCount members were random values (whatever was in memory/RAM
at the time), when supplied to `generateRandomData()`. My first
attempt at running caused a crash, because it was trying to allocate
a few billion charts :D

I have added -Wall, -Wextra and -Werror to Cmake project. Note that
these assume you are using GCC or Clang. MSVC has a different format.

-Wall: Raise *all* compile warnings
  - It's not all possible warnings, but basically every sensible one
    we need thanks to all the crap in the language, like init order :'(
- Wextra: Enable even more warnings not covered by Wall. Just go with
  these. Every professional dev outfit will have enabled.
- Werror: Treat all warnings as errors, and fail compilation.
  - If there is a warning, it's most likely worrying about. Feel free
    to remove this if you are getting stuck on a warning while learning.

I've grouped and separated  data members from functions in the
class header, purely as a stylistic thing. I've never worked at a place
that didn't do this as the norm. Feel free to order your groupings
how you see fit, but please group them :)